server/Python_Core/requirements_unified.txt
# ========================================
# UNIFIED PYTHON SERVICE REQUIREMENTS
# For: Text Analysis, Body Language Analysis, and Speech Analysis
# ========================================

# Web Framework
fastapi==0.116.2
uvicorn==0.35.0
python-multipart==0.0.20
starlette==0.48.0

# HTTP and Utilities
requests==2.32.5
python-dotenv==1.0.0
click==8.2.1
pydantic==2.11.9
pydantic_core==2.33.2

# Audio Processing (Speech Analysis)
librosa==0.10.1
soundfile==0.12.1
noisereduce==3.0.0
sounddevice==0.5.2
praat-parselmouth==0.4.3
resampy>=0.4.2

# Video Processing (Body Language Analysis)
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
mediapipe==0.10.21

# Natural Language Processing (All modules)
nltk==3.9.2
phonetics==1.0.5
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.2.5
wasabi==1.1.3
weasel==0.4.1
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85

# Speech Recognition (Speech Analysis)
openai-whisper==20231117
faster-whisper==1.0.3

# Machine Learning
scikit-learn==1.7.2
transformers==4.35.2
torch>=2.2,<2.3
sentencepiece==0.2.1

# Data Science and Numerical Computing
numpy==1.26.4
scipy==1.16.2
pandas
matplotlib==3.10.6
joblib==1.5.2

# Text Analysis Tools
language_tool_python==2.9.4
textstat==0.7.10
pyphen==0.17.2

# Language Detection
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1

# Deep Learning (JAX for some models)
jax==0.7.1
jaxlib==0.7.1
ml_dtypes==0.5.3
opt_einsum==3.4.0
flatbuffers==25.2.10

# Data Structures
blis==0.7.11
catalogue==2.0.10
cymem==2.0.11
murmurhash==1.0.13
preshed==3.0.10

# Utilities
tqdm==4.67.1
typer==0.19.2
rich==14.1.0
colorama==0.4.6
smart_open==7.3.1
cloudpathlib==0.23.0
shellingham==1.5.4

# Web and Data Formats
certifi==2025.10.5
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
h11==0.16.0
anyio==4.10.0
sniffio==1.3.1

# Template and Markup
Jinja2==3.1.6
MarkupSafe==3.0.3
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2

# Plotting
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.0
kiwisolver==1.4.9
pillow==11.3.0

# Configuration and Parsing
confection==0.1.5
packaging==25.0
pyparsing==3.2.4
toml==0.10.2
regex==2025.9.18

# System and Concurrency
psutil==7.1.0
threadpoolctl==3.6.0
wrapt==1.17.3

# Date/Time
python-dateutil==2.9.0.post0

# Legacy Support
six==1.17.0
typing_extensions==4.15.0
typing-inspection==0.4.1

# Development
setuptools==80.9.0
